+++ app/Http/Controllers/RemoveCropsFromFieldController.php
@@ -7,26 +7,19 @@ namespace App\Http\Controllers;
 use App\Models\Field;
 use App\Models\Crop;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Symfony\Component\Console\Input\Input;
 
 class RemoveCropsFromFieldController extends Controller
 {
-    /**
-     * Handle the incoming request.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
     public function __invoke(Request $request, Field $field, Crop $crop)
     {
-        $field = Field::find($request->segment(2));
-        $crop_id = $request->segment(3);
-        $field->crops()->detach([$crop_id]);
+        if (Auth::user()) {
+            $field = Field::find($request->segment(2));
+            $crop_id = $request->segment(3);
+            $field->crops()->detach([$crop_id]);
 
-
-        // Ta id:na från URL:en och ta bort croppen från rätt field (från databasen). 
-        // Både crop-id och fields-id hänger mer i URL:en. 
-
-        return redirect('/dashboard');
+            return redirect('/dashboard');
+        }
     }
 }
